+++ src/page-objects/administration/RuleCreate.ts
@@ -8,6 +8,9 @@ export class RuleCreate implements PageObject {
     public readonly header: Locator;
     public readonly nameInput: Locator;
     public readonly priorityInput: Locator;
+    public readonly descriptionInput: Locator;
+    public readonly typeItem: Locator;
+    public readonly tagItem: Locator;
     public readonly conditionTypeSelectionInput: Locator;
     public readonly conditionValueSelectionInput: Locator;
     public readonly filtersResultPopoverSelectionList: Locator;
@@ -23,6 +26,9 @@ export class RuleCreate implements PageObject {
         this.header = page.locator('.smart-bar__header');
         this.nameInput = page.getByLabel(translate('administration:rule:fields.name'));
         this.priorityInput = page.getByLabel(translate('administration:rule:fields.priority'));
+        this.descriptionInput = page.getByLabel('Description');
+        this.typeItem = page.locator('.sw-settings-rule-detail__type-field').locator('.sw-select-selection-list__item');
+        this.tagItem = page.locator('.sw-settings-rule-detail__tags-field').locator('.sw-select-selection-list__item');
         this.conditionTypeSelectionInput = page.locator('.sw-condition-type-select').locator('.sw-single-select__selection');
         this.conditionValueSelectionInput = page.locator('.sw-condition__value-content').locator('.sw-entity-single-select__selection');
         this.filtersResultPopoverSelectionList = page.locator('.sw-select-result-list__content').getByRole('listitem');