+++ src/page-objects/storefront/CheckoutCart.ts
@@ -12,6 +12,9 @@ export class CheckoutCart implements PageObject {
     public readonly cartLineItemImages: Locator;
     public readonly unitPriceInfo: Locator;
     public readonly cartQuantityNumber: Locator;
+    public readonly productNameLabel:(productName: string) => Locator;
+    public readonly productNumberLabel:(productNumber: string) => Locator;
+    public readonly productDeliveryDateLabel: Locator;
 
     public readonly page: Page;
 
@@ -20,12 +23,15 @@ export class CheckoutCart implements PageObject {
         this.headline = page.getByRole('heading', { name: translate('storefront:checkout:cart.shoppingCart') });
         this.goToCheckoutButton = page.getByRole('link', { name: translate('storefront:checkout:cart.goToCheckout') });
         this.enterPromoInput = page.getByLabel(translate('storefront:checkout:cart.promoCode'));
-        this.grandTotalPrice = page.locator('dt:has-text("Grand total") + dd:visible');
+        this.grandTotalPrice = page.locator(`dt:has-text("${translate('storefront:checkout:finish.grandTotal')}") + dd:visible`);
         this.emptyCartAlert = page.getByText(translate('storefront:checkout:cart.emptyCart'));
         this.stockReachedAlert = page.getByText(translate('storefront:checkout:cart.stockReached'));
         this.cartLineItemImages = page.locator('.line-item-img-link');
         this.unitPriceInfo = page.locator('.line-item-unit-price-value');
         this.cartQuantityNumber = page.locator('input[name="quantity"]');
+        this.productNameLabel = (productName: string) => page.getByRole('link', { name: productName });
+        this.productNumberLabel = (productNumber: string) => page.getByText(productNumber);
+        this.productDeliveryDateLabel = page.locator('.line-item-delivery-date');
     }
 
     url() {